StackOverflow.jl Documentation

StackOverflow.jl Documentation

Functions

Index

Functions

StackOverflow.getrecentquestionsfortag — Function

getrecentquestionsfortag(;tag::String = "Julia", site::String = "stackoverflow", order::String = "desc",
                                sort::String = "activity", todate::String = "", fromdate::String = "")

Will return the 30 StackOverflow questions given a certain tag and sort parameter.

Possible order values: "desc" or "asc" Possible sort values: "activity", "votes", "creation", "hot", "week", or "month"

StackOverflow.checkmyprivileges — Function

checkmyprivileges(username::String)

Takes in a username and returns the privileges said user has.

StackOverflow.getprivileges — Function

getprivileges()

Prints out the various privilege levels in Stack Overflow and the rep required to have said privileges.

Missing docstring.

Missing docstring for getquestions. Check Documenter's build log for details.

StackOverflow.getrecentbadgerecipients — Function

getrecentbadgerecipients()

Returns recently awarded badges in the system.

StackOverflow.getsiteinfo — Function

getsiteinfo(;site::String = "stackoverflow")

Will return a collection of statistics about the site (StackOverflow).

StackOverflow.searcherror — Function

searcherror(intitle::String)

Adding the ability to search for an error through StackOverflow.jl while in the Julia Terminal.

StackOverflow.searchtag — Function

searchtag(intitle::String)

Adding the ability to search for an tag through StackOverflow.jl while in the Julia Terminal.

StackOverflow.getfeaturedquestions — Function

getfeaturedquestions(page::Int = 1, pagesize::Int = 1, fromdate::String = "", todate::String = "",
                        order::String = "desc", sort::String = "activity", tags::String = "Julia",
                        site::String = "stackoverflow")

Will return bountied questions given a tag. pagesize is the number of questions it will return, page is the page on which it takes questions from, and returns pagesize number of questions.

Possible order values: "desc" or "asc" Possible sort values: "activity", "votes", "creation", "hot", "week", or "month"

StackOverflow.getopenquestions — Function

getopenquestions(page::Int = 1, pagesize::Int = 1, fromdate::String = "", todate::String = "",
                        order::String = "desc", sort::String = "activity", tags::String = "Julia",
                        site::String = "stackoverflow")

Will return questions that have no marked answer given a tag. pagesize is the number of questions it will return, page is the page on which it takes questions from, and returns pagesize number of questions.

Possible order values: "desc" or "asc" Possible sort values: "activity", "votes", "creation", "hot", "week", or "month"

StackOverflow.getmoderators — Function

getmoderators(page::Int = 1, pagesize::Int = 1, fromdate::String = "", todate::String = "",
            order::String = "desc", site::String = "stackoverflow", sort::String = "creation")

Returns moderators on a site. Returns pagesize moderators on page.

Possible order values: "desc" or "asc" Possible sort values: "reputation", "creation", "name", "modified"

StackOverflow.getnoanswerquestions — Function

getnoanswerquestions(page::Int = 1, pagesize::Int = 1, fromdate::String = "", todate::String = "",
                        order::String = "desc", sort::String = "activity", tags::String = "Julia",
                        site::String = "stackoverflow")

Will return questions with zero answers given a tag. pagesize is the number of questions it will return, page is the page on which it takes questions from, and returns pagesize number of questions.

Possible order values: "desc" or "asc" Possible sort values: "activity", "votes", "creation", "hot", "week", or "month"

Index

StackOverflow.checkmyprivileges

StackOverflow.getfeaturedquestions

StackOverflow.getmoderators

StackOverflow.getnoanswerquestions

StackOverflow.getopenquestions

StackOverflow.getprivileges

StackOverflow.getrecentbadgerecipients

StackOverflow.getrecentquestionsfortag

StackOverflow.getsiteinfo

StackOverflow.searcherror

StackOverflow.searchtag